+++ src/actions.ts
@@ -2,7 +2,6 @@ import type { App, Editor, EditorPosition, EditorSelection } from 'obsidian';
 import {
   CASE,
   DIRECTION,
-  LOWERCASE_ARTICLES,
   MATCHING_BRACKETS,
   MATCHING_QUOTES,
   MATCHING_QUOTES_BRACKETS,
@@ -18,8 +17,10 @@ import {
   getLeadingWhitespace,
   getLineEndPos,
   getLineStartPos,
+  getNextCase,
   getSearchText,
   getSelectionBoundaries,
+  toTitleCase,
   wordRangeAtPos,
 } from './utils';
 
@@ -390,44 +391,6 @@ export const transformCase = (
   editor.replaceRange(replacementText, from, to);
 
   return selection;
-};
-
-export const toTitleCase = (selectedText: string) => {
-  // use capture group to join with the same separator used to split
-  return selectedText
-      .split(/(\s+)/)
-      .map((word, index, allWords) => {
-        if (
-            index > 0 &&
-            index < allWords.length - 1 &&
-            LOWERCASE_ARTICLES.includes(word.toLowerCase())
-        ) {
-          return word.toLowerCase();
-        }
-        return word.charAt(0).toUpperCase() + word.substring(1).toLowerCase();
-      })
-      .join('');
-};
-
-export const getNextCase = (selectedText: string): string => {
-  const textUpper = selectedText.toUpperCase();
-  const textLower = selectedText.toLowerCase();
-  const textTitle = toTitleCase(selectedText);
-
-  switch (selectedText) {
-    case textUpper: {
-      return textLower;
-    }
-    case textLower: {
-      return textTitle;
-    }
-    case textTitle: {
-      return textUpper;
-    }
-    default: {
-      return textUpper;
-    }
-  }
 };
 
 const expandSelection = ({
