+++ src/main.ts
@@ -292,13 +292,13 @@ export default class CodeEditorShortcuts extends Plugin {
     });
 
     this.addCommand({
-      id: 'transformToggleCase',
+      id: 'toggleCase',
       name: 'Toggle case of selection',
       editorCallback: (editor) =>
-          withMultipleSelections(editor, transformCase, {
-            ...defaultMultipleSelectionOptions,
-            args: CASE.NEXT,
-          }),
+        withMultipleSelections(editor, transformCase, {
+          ...defaultMultipleSelectionOptions,
+          args: CASE.NEXT,
+        }),
     });
 
     this.addCommand({
